Make content text bottom margin handling more explicit
There are different cases where we want to add a margin to the bottom
of the content text to give the scroller some extra padding:

- Prevent content text from overlapping scroll indicator on phone

- Prevent content text from overlapping player control/info box

If the scroller is clipped though (for exaple above the waveform), we
want to be able to remove the margin.

Make content text margin explicitly configurable in the `PageText`
component instead of hard coding special cases for the audio and video
pages in the CSS.

Leaving room for the phone scroll indicator has to be the default
behavior since the existing non-React page types do not apply
additional classes to the `contentText` element yet.