Add specs for permissions config
Whenever we add policy specs, it is a good idea to begin reflecting
admin permissions in the policy. As long as we use CanCan or similar,
this is redundant, because our use of :manage lets admins do even
custom verbs, not only CRUD.

As soon as we switch, policies without admin permissions would not
provide what the abilities currently do for us, and without
test-covering admin permissions very thoroughly in addition to testing
permissions, we would face a big investment of effort.